Flytter validering av beregning.
