+++ src/main/kotlin/no/nav/familie/ba/sak/kjerne/behandling/domene/Behandling.kt
@@ -126,6 +126,14 @@ data class Behandling(
         return this
     }
 
+    fun skalRettFraBehandlingsresultatTilIverksetting(): Boolean {
+        return when {
+            skalBehandlesAutomatisk && erOmregning() && resultat == BehandlingResultat.FORTSATT_INNVILGET -> true
+            skalBehandlesAutomatisk && resultat == BehandlingResultat.INNVILGET -> true
+            else -> false
+        }
+    }
+
     private fun leggTilStegOmDetIkkeFinnesFraFør(steg: StegType) {
         if (behandlingStegTilstand.none { it.behandlingSteg == steg }) {
             behandlingStegTilstand.add(
@@ -165,9 +173,11 @@ data class Behandling(
 
     fun erKlage(): Boolean = this.opprettetÅrsak == BehandlingÅrsak.KLAGE
 
-    fun erMigrering() = type == BehandlingType.MIGRERING_FRA_INFOTRYGD || type == BehandlingType.MIGRERING_FRA_INFOTRYGD_OPPHØRT
+    fun erMigrering() =
+        type == BehandlingType.MIGRERING_FRA_INFOTRYGD || type == BehandlingType.MIGRERING_FRA_INFOTRYGD_OPPHØRT
 
-    fun erOmregning(): Boolean = this.opprettetÅrsak == BehandlingÅrsak.OMREGNING_6ÅR || this.opprettetÅrsak == BehandlingÅrsak.OMREGNING_18ÅR
+    fun erOmregning(): Boolean =
+        this.opprettetÅrsak == BehandlingÅrsak.OMREGNING_6ÅR || this.opprettetÅrsak == BehandlingÅrsak.OMREGNING_18ÅR
 
     fun hentYtelseTypeTilVilkår(): YtelseType = when (underkategori) {
         BehandlingUnderkategori.UTVIDET -> YtelseType.UTVIDET_BARNETRYGD
